+++ .gitignore
@@ -509,3 +509,4 @@ override.tf.json
 # Ignore CLI configuration files
 .terraformrc
 terraform.rc
+.DS_Store
