+++ fault-quarantine-module/Dockerfile
@@ -35,7 +35,7 @@ RUN --mount=type=cache,target=/go/pkg/mod \
     go mod download
 
 RUN --mount=type=cache,target=/go/pkg/mod \
-    go build -trimpath -o faultquarantine cmd/faultquarantine.go
+    go build -trimpath -o faultquarantine main.go
 
 FROM ${IMG}:${TAG}
