+++ fault-quarantine-module/main.go
@@ -20,38 +20,63 @@ import (
 	"fmt"
 	"net/http"
 	"os"
+	"os/signal"
 	"path/filepath"
 	"strconv"
+	"syscall"
+	"time"
 
-	config "github.com/nvidia/nvsentinel/health-events-analyzer/pkg/config"
-	"github.com/nvidia/nvsentinel/health-events-analyzer/pkg/publisher"
-	"github.com/nvidia/nvsentinel/health-events-analyzer/pkg/reconciler"
-	pb "github.com/nvidia/nvsentinel/platform-connectors/pkg/protos"
+	"github.com/nvidia/nvsentinel/fault-quarantine-module/pkg/config"
+	"github.com/nvidia/nvsentinel/fault-quarantine-module/pkg/reconciler"
 	"github.com/nvidia/nvsentinel/store-client-sdk/pkg/storewatcher"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
-	"k8s.io/klog"
+	"k8s.io/klog/v2"
 )
 
-//nolint:cyclop // todo
+var (
+	version = "dev"
+	commit  = "none"
+	date    = "unknown"
+)
+
+// nolint: cyclop //fix this as part of NGCC-21793
 func main() {
-	ctx := context.Background()
+	klog.Infof("Fault Quarantine Module - version: %s, commit: %s, date: %s", version, commit, date)
+
+	// Create a context that gets cancelled on OS interrupt signals
+	ctx, stop := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGTERM)
+	defer stop() // Ensure the signal listener is cleaned up
 
 	var metricsPort = flag.String("metrics-port", "2112", "port to expose Prometheus metrics on")
 
-	var socket = flag.String("socket", "unix:///var/run/nvsentinel.sock", "unix domain socket")
-
 	var mongoClientCertMountPath = flag.String("mongo-client-cert-mount-path", "/etc/ssl/mongo-client",
 		"path where the mongodb client cert is mounted")
+
+	var kubeconfigPath = flag.String("kubeconfig-path", "", "path to kubeconfig file")
+
+	var dryRun = flag.Bool("dry-run", false, "flag to run node drainer module in dry-run mode")
+
+	var circuitBreakerPercentage = flag.Int("circuit-breaker-percentage",
+		50, "percentage of nodes to cordon before tripping the circuit breaker")
+
+	var circuitBreakerDuration = flag.Duration("circuit-breaker-duration",
+		5*time.Minute, "duration of the circuit breaker window")
+
+	var circuitBreakerEnabled = flag.Bool("circuit-breaker-enabled", true,
+		"enable or disable fault quarantine circuit breaker")
 
 	// Initialize klog flags to allow command-line control (e.g., -v=3)
 	klog.InitFlags(nil)
 
 	defer klog.Flush()
 	flag.Parse()
+
+	namespace := os.Getenv("POD_NAMESPACE")
+	if namespace == "" {
+		klog.Fatalf("POD_NAMESPACE is not provided")
+	}
 
 	mongoURI := os.Getenv("MONGODB_URI")
 	if mongoURI == "" {
@@ -98,6 +123,12 @@ func main() {
 		klog.Fatalf("invalid CA_CERT_READ_INTERVAL_SECONDS: %v", err)
 	}
 
+	unprocessedEventsMetricUpdateIntervalSeconds, err :=
+		getEnvAsInt("UNPROCESSED_EVENTS_METRIC_UPDATE_INTERVAL_SECONDS", 25)
+	if err != nil {
+		klog.Fatalf("invalid UNPROCESSED_EVENTS_METRIC_UPDATE_INTERVAL_SECONDS: %v", err)
+	}
+
 	mongoConfig := storewatcher.MongoDBConfig{
 		URI:        mongoURI,
 		Database:   mongoDatabase,
@@ -114,48 +145,55 @@ func main() {
 	}
 
 	tokenConfig := storewatcher.TokenConfig{
-		ClientName:      "health-events-analyzer",
+		ClientName:      "fault-quarantine-module",
 		TokenDatabase:   tokenDatabase,
 		TokenCollection: tokenCollection,
 	}
 
 	pipeline := mongo.Pipeline{
-		bson.D{
-			{Key: "$match", Value: bson.D{
-				{Key: "operationType", Value: "insert"},
-				{Key: "fullDocument.healthevent.isfatal", Value: false},
-				{Key: "fullDocument.healthevent.ishealthy", Value: false},
-			}},
-		},
+		{{Key: "$match", Value: bson.D{{Key: "operationType", Value: bson.D{{Key: "$in", Value: bson.A{"insert"}}}}}}},
 	}
 
-	var opts []grpc.DialOption
-	opts = append(opts, grpc.WithTransportCredentials(insecure.NewCredentials()))
-
-	conn, err := grpc.NewClient(*socket, opts...)
+	tomlCfg, err := config.LoadTomlConfig("/etc/config/config.toml")
 	if err != nil {
-		panic(err)
+		klog.Fatalf("error while loading the toml config: %v", err)
 	}
-	defer conn.Close()
 
-	platformConnectorClient := pb.NewPlatformConnectorClient(conn)
-	publisher := publisher.NewPublisher(platformConnectorClient)
+	if *dryRun {
+		klog.Info("Running in dry-run mode")
+	}
 
-	// Parse the TOML content
-	tomlConfig, err := config.LoadTomlConfig("/etc/config/config.toml")
+	// Initialize the k8s client
+	k8sClient, err := reconciler.NewFaultQuarantineClient(*kubeconfigPath, *dryRun)
 	if err != nil {
-		klog.Fatalf("Failed to load config file: %v", err)
+		klog.Fatalf("error while initializing kubernetes client: %v", err)
 	}
 
-	reconcilerCfg := reconciler.HealthEventsAnalyzerReconcilerConfig{
+	klog.Info("Successfully initialized k8sclient")
+
+	reconcilerCfg := reconciler.ReconcilerConfig{
+		TomlConfig:                       *tomlCfg,
 		MongoHealthEventCollectionConfig: mongoConfig,
 		TokenConfig:                      tokenConfig,
 		MongoPipeline:                    pipeline,
-		HealthEventsAnalyzerRules:        tomlConfig,
-		Publisher:                        publisher,
+		K8sClient:                        k8sClient,
+		DryRun:                           *dryRun,
+		CircuitBreakerEnabled:            *circuitBreakerEnabled,
+		UnprocessedEventsMetricUpdateInterval: time.Duration(unprocessedEventsMetricUpdateIntervalSeconds) *
+			time.Second,
+		CircuitBreaker: reconciler.CircuitBreakerConfig{
+			Namespace:  namespace,
+			Name:       "fault-quarantine-circuit-breaker",
+			Percentage: *circuitBreakerPercentage,
+			Duration:   *circuitBreakerDuration,
+		},
 	}
 
-	reconciler := reconciler.NewReconciler(reconcilerCfg)
+	// Create the work signal channel (buffered channel acting as semaphore)
+	workSignal := make(chan struct{}, 1) // Buffer size 1 is usually sufficient
+
+	// Pass the workSignal channel to the Reconciler
+	reconciler := reconciler.NewReconciler(ctx, reconcilerCfg, workSignal)
 
 	go func() {
 		http.Handle("/metrics", promhttp.Handler())
