+++ health-monitors/syslog-health-monitor/main.go
